% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_data.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster data with fuzzy c-means}
\usage{
cluster_data(distM, data, ncl)
}
\arguments{
\item{distM}{a distance matrix. Computed using compute_dist_0, compute_dist_1 or compute_dist_2.}

\item{data}{a dataframe. Includes at least the key column 'filename'.}

\item{ncl}{a numeric scalar. The number of clusters to be specified.}
}
\value{
a dataframe. Includes the fuzzy clustering results.
}
\description{
Cluster data with fuzzy c-means
}
\examples{
distM = read.csv('~/myRpacks/morphotype/inst/extdata/distM.csv', row.names = 1)
data = read.csv('~/myRpacks/morphotype/inst/extdata/data_0D.csv')
data = as.data.frame(data, stringsAsFactors = FALSE)
data = data[,c("filename","TV", "Type")]
cluster_data(distM = distM, data = data, ncl = 3)
}
\author{
Danai Kafetzaki
}
