% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine two sources of morphological data}
\usage{
combine_data(data1, data2, PCs)
}
\arguments{
\item{data1}{a distance matrix. Computed using compute_dist_0, compute_dist_1 or compute_dist_2.}

\item{data2}{a dataframe. Includes the key column 'filename' and the columns with the morphological information.}

\item{PCs}{a numeric scalar. The number of principal components to be retained. The default is to select the components that explain more than 1\% of the variance.}
}
\value{
A list with the following items:
\itemize{
\item df_eig.val - a dataframe. Includes the eigenvalues and variances of the principal dimensions.
\item df_combined - a dataframe. The new dataframe which combines the two input data sources.
}
}
\description{
Combine two sources of morphological data
}
\examples{
data1 = read.csv('~/myRpacks/morphotype/inst/extdata/data_1D.csv')
data2 = read.csv('~/myRpacks/morphotype/inst/extdata/data_0D.csv')
data2 = as.data.frame(data2, stringsAsFactors = FALSE)
data2 = data2[,c("filename","sherd_rim.diameter","Below.Rim_incl_sin_min")]
combine_data(data1 = data1, data2 = data2)
combine_data(data1 = data1, data2 = data2, PCs = 3)
}
\author{
Danai Kafetzaki
}
