% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dist_0.R
\name{compute_dist_0}
\alias{compute_dist_0}
\title{Compute distance between point measures}
\usage{
compute_dist_0(data, fun, p_mink = 4)
}
\arguments{
\item{data}{a dataframe. The first column includes the filename of the profile drawing which serves as key. The following columns include numeric data which should be used for computing the distance matrix.}

\item{fun}{a character vector. The distance method that should be used. This is one of the following: "euclidean", "manhattan", "canberra", "minkowski", "chebysev".}

\item{p_mink}{a numeric scalar. The order that should be used for the Minkowski distance. This parameter is optional. In case the Minkowski distance is computed, p_mink takes the default value of 4.}
}
\value{
A list with the following items:
\itemize{
\item data_normalised - a dataframe. The normalised version of the input data.
\item dist_matrix - a 2D matrix. The distance matrix of the normalised data.
}
}
\description{
Compute distance between point measures
}
\examples{
data = read.csv('~/myRpacks/morphotype/inst/extdata/data_0D.csv')
data = as.data.frame(data, stringsAsFactors = FALSE)
data = data[,c(1:28, 31:32, 37, 47:58, 62:65)]
compute_dist_0(data = data, fun = "manhattan")
compute_dist_0(data = data, fun = "minkowski", p_mink = 4)
}
\author{
Danai Kafetzaki
}
