% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dist_1.R
\name{compute_dist_1}
\alias{compute_dist_1}
\title{Compute distance between one-dimensional outlines}
\usage{
compute_dist_1(data, fun, p_mink = 2)
}
\arguments{
\item{data}{a dataframe. The first column of the dataframe includes the index of the outline. The rest of the columns include the scaled 1D outline of each profile under study.}

\item{fun}{a character vector. The distance method that should be used, which is one of the following: "manhattan", "chebysev", "canberra", "minkowski", "euclidean", "pearson", "spearman", "kendall", "kst".}

\item{p_mink}{a numeric scalar. The order that should be used for the Minkowski distance. This parameter is optional. In case the Minkowski distance is computed, p_mink takes the default value of 4.}
}
\value{
A list with the following items:
\itemize{
\item data_normalised - a dataframe. The normalised version of the input data.
\item dist_matrix - a 2D matrix. The distance matrix of the normalised data.
}
}
\description{
Compute distance between one-dimensional outlines
}
\examples{
data = read.csv('~/myRpacks/morphotype/inst/extdata/data_1D.csv')
compute_dist_1(data = data, fun = "manhattan")
compute_dist_1(data = data, fun = "minkowski", p_mink = 4)
}
\author{
Danai Kafetzaki
}
