% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dist_2.R
\name{compute_dist_2}
\alias{compute_dist_2}
\title{Compute distance between two-dimensional outlines}
\usage{
compute_dist_2(data, fun, pS = 10, pD = 10)
}
\arguments{
\item{data}{a dataframe.}

\item{fun}{a character vector. The distance method that should be used, which is one of the following: "dtw", "jaccard".}

\item{pS}{a numeric scalar.}

\item{pD}{a numeric scalar.}
}
\value{
A list with the following items:
\itemize{
\item data_normalised - a dataframe. The normalised version of the input data.
\item dist_matrix - a 2D matrix. The distance matrix of the normalised data.
}
}
\description{
Compute distance between two-dimensional outlines
}
\examples{
load('~/myRpacks/morphotype/data/data_2D.RData')
compute_dist_2(data = mat_circumference_scaled[1:2], fun = "edit")
}
\author{
Danai Kafetzaki
}
