% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_GAM.R
\name{compute_GAM}
\alias{compute_GAM}
\title{Computes GAM representation of profile wall}
\usage{
compute_GAM(df_profiles_2D_ScH)
}
\arguments{
\item{df_profiles_2D_ScH}{a list of numeric matrices. Each numeric matrix contains the data for the rescaled profile, as a result of function get_data_GAM().}
}
\value{
A list with the following items:
\itemize{
\item mat_spl_FP - a list of a 2D vector for each profile in the dataset. The y axis is the GAM vector for each profile and the x axis is the index sequence which by construction has equal length for all profiles in the dataset.
}
}
\description{
Computes GAM representation of profile wall
}
\examples{
load(file = "~/myRpacks/morphotype/data/profile_ScH_1C100_20230405.RData")
compute_GAM(df_profiles_2D_ScH = mat_cropped_profile_ScH)
}
\author{
Danai Kafetzaki
}
