% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_circ_bop.R
\name{crop_circ_bop}
\alias{crop_circ_bop}
\title{Crops circumference at bottom outer point}
\usage{
crop_circ_bop(outer.line, inner.line, marSc)
}
\arguments{
\item{outer.line}{a numeric 1D vector. The outer profile line.}

\item{inner.line}{a numeric 1D vector. The inner profile line.}

\item{marSc}{a numeric scalar. The amount of pixels vertically after the outer wall of the base that will be included in the result.}
}
\value{
A list with the following items:
\itemize{
\item mati_scaled - a dataframe with numeric attributes. The x and y position of the points belonging to the circumference.
\item thrP - a numeric scalar. The x index value of the custom cut of the profile base.
\item indx - a numeric 1D vector. The y indexes with a value in the custom cut of the profile base.
}
}
\description{
Crops circumference at bottom outer point
}
\examples{
crop_circ_bop(outer.line = m19$outer.line, inner.line = m19$inner.line, marSc = 5)
}
\author{
Danai Kafetzaki
}
