% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_prof_bip.R
\name{crop_prof_bip}
\alias{crop_prof_bip}
\title{Crops profile at bottom inner point}
\usage{
crop_prof_bip(profile.wall, thrP, indx)
}
\arguments{
\item{profile.wall}{a numeric 2D matrix. The binary matrix of the profile wall.}

\item{thrP}{a numeric scalar. The x index value of the custom cut of the profile base.}

\item{indx}{a numeric 1D vector. The y indexes with a value in the custom cut of the profile base.}
}
\value{
A list with the following items:
\itemize{
\item df_img_check - Image object, as a result of EBImage formal class Image.
}
}
\description{
Crops profile at bottom inner point
}
\examples{
crop_prof_bip(profile.wall = m19$profile.wall, thrP = 124, indx = c(1:12))
}
\author{
Danai Kafetzaki
}
