% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_img_check.R
\docType{data}
\name{df_img_check}
\alias{df_img_check}
\title{Output of ....}
\format{
A list of 3 items:
\describe{
\item{Item1}{description}
\item{Item2}{description}
\item{Item3}{description}
}
}
\source{
\url{https://www.dataURL.com}
}
\usage{
df_img_check
}
\description{
More about Output of ...
}
\keyword{datasets}
