% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\docType{data}
\name{filename}
\alias{filename}
\title{Output of ....}
\format{
A list of 3 items:
\describe{
\item{Item1}{description}
\item{Item2}{description}
\item{Item3}{description}
}
}
\source{
\url{https://www.dataURL.com}
}
\usage{
filename
}
\description{
More about Output of ...
}
\keyword{datasets}
