% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angle_rim.R
\name{get_angle_rim}
\alias{get_angle_rim}
\title{Compute the angle between two points of the profile}
\usage{
get_angle_rim(profile.wall, m9, m28)
}
\arguments{
\item{profile.wall}{a numeric 2D matrix. The binary matrix of the profile wall.}

\item{m9}{a numeric scalar. The height of the profile in pixels.}

\item{m28}{a numeric scalar. The height of the profile segment in pixels.}
}
\value{
A list with the following items:
\itemize{
\item Rim_incl_min - Sine value for the inclination measured at the most outer wall coordinates of the bottom and top of the profile segment.
\item Rim_incl_max - Sine value for the inclination measured at the most inner wall coordinates of the bottom and top of the profile segment.
\item Rim_incl_mean - Sine value for the inclination measured at the middle coordinates of the bottom and top of the profile segment.
\item Rim_incl_min_sign - Sign of the Rim_incl_min, coded as -1 for negative, 1 for positive and 0 for 90 degrees angle.
\item Rim_incl_max_sign - Sign of the Rim_incl_max, coded as -1 for negative, 1 for positive and 0 for 90 degrees angle.
\item Rim_incl_mean_sign - Sign of the Rim_incl_mean, coded as -1 for negative, 1 for positive and 0 for 90 degrees angle.
\item Rim_incl_distance_min - Actual length in pixels of the conceivable hypotenuse for the triangle measuring Rim_incl_min.
\item Rim_incl_distance_max - Actual length in pixels of the conceivable hypotenuse for the triangle measuring Rim_incl_max.
\item Rim_incl_distance_mean - Actual length in pixels of the conceivable hypotenuse for the triangle measuring Rim_incl_mean.

}
}
\description{
Compute the angle between two points of the profile
}
\examples{
get_angle_rim(profile.wall = m19$profile.wall, m9 = m9, m28 = 67)
}
\author{
Danai Kafetzaki
}
