% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_wall.R
\name{get_average_wall}
\alias{get_average_wall}
\title{Compute the wall thickness across the profile height and relevant statistics}
\usage{
get_average_wall(outer.line, inner.line)
}
\arguments{
\item{outer.line}{a numeric 1D vector. The outer profile line.}

\item{inner.line}{a numeric 1D vector. The inner profile line.}
}
\value{
A list with the following items:
\itemize{
\item all_WT - a numeric 1D vector including the wall thickness in pixels across the height of the profile wall.
\item mean.WT - the average of all_WT.
\item median.WT - the median of all_WT.
\item sd.WT - the standard deviation of all_WT.
\item min.WT - the minimum of all_WT.
\item max.WT - the maximum of all_WT.
}
}
\description{
Compute the wall thickness across the profile height and relevant statistics
}
\examples{
get_average_wall(outer.line = m19$outer.line, inner.line = m19$inner.line)
}
\author{
Danai Kafetzaki
}
