% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bottom_diameter.R
\name{get_bottom_diameter}
\alias{get_bottom_diameter}
\title{Compute the diameter at the base of the profile}
\usage{
get_bottom_diameter(xx)
}
\arguments{
\item{xx}{a dataframe. Melted profile matrix as a result of get_transforms.}
}
\value{
a numeric scalar. The diameter at the bottom of the profile. In case the profile is recognized as full profile, the value is the diameter at the profile base.
}
\description{
Compute the diameter at the base of the profile
}
\examples{
get_bottom_diameter(xx = m10)
}
\author{
Danai Kafetzaki
}
