% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bottomline.R
\name{get_bottomline}
\alias{get_bottomline}
\title{Compute the coordinate of the profile's bottom line}
\usage{
get_bottomline(image_denoised, Y_min_i, margin)
}
\arguments{
\item{image_denoised}{a numeric 2D matrix. The 2D binary matrix of the input image.}

\item{Y_min_i}{a numeric 1D vector. The y index for each y including data across the image height.}

\item{margin}{a numeric scalar. The number of pixels checked for correction in the calculation.}
}
\value{
a numeric scalar. The y coordinate of the most bottom point of the profile.
}
\description{
Compute the coordinate of the profile's bottom line
}
\examples{
get_bottomline(image_denoised = image_denoised, Y_min_i = m2$Y_min_i, margin = 6)
}
\author{
Danai Kafetzaki
}
