% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_circumference.R
\name{get_circumference}
\alias{get_circumference}
\title{Compute the circumference of the profile}
\usage{
get_circumference(outer.line, inner.line)
}
\arguments{
\item{outer.line}{a numeric 1D vector. The outer profile line.}

\item{inner.line}{a numeric 1D vector. The inner profile line.}
}
\value{
A list with the following items:
\itemize{
\item bottom_width - a numeric scalar. The width of the bottom in the rim profile.
\item lip_width - a numeric scalar. The width of the top in the rim profile.
\item data_circumference - a dataframe with numeric attributes. The x and y position of the points belonging to the circumference.
}
}
\description{
Compute the circumference of the profile
}
\examples{
get_circumference(outer.line = m53$outer.rim.line.scaled,
                  inner.line = m53$inner.rim.line.scaled)
}
\author{
Danai Kafetzaki
}
