% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curl.R
\name{get_curl}
\alias{get_curl}
\title{Compute the curl of the profile segment}
\usage{
get_curl(perip.original, m28)
}
\arguments{
\item{perip.original}{a dataframe with one named numeric vector. The periphery of the profile as a result of the function get_vector_df.}

\item{m28}{a numeric scalar. The height from the top of the profile of the point defining the considered segment.}
}
\value{
a numeric scalar. The curl value for the profile segment.
}
\description{
Compute the curl of the profile segment
}
\examples{
get_curl(perip.original = m58$perip.original, m28 = 67)
}
\author{
Danai Kafetzaki
}
