% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curvature.R
\name{get_curvature}
\alias{get_curvature}
\title{Compute the curvature and the slope for a given outline.}
\usage{
get_curvature(outline)
}
\arguments{
\item{outline}{a numeric 1D vector. The outline of the profile.}
}
\value{
A list with the following items:
\itemize{
\item curvf - a numeric 1D vector. The curvature representation of the given profile.
\item slop - a numeric 1D vector. The tangent representation of the given profile.
}
}
\description{
Compute the curvature and the slope for a given outline.
}
\examples{
get_curvature(outline = m19$outer.line)
}
\author{
Danai Kafetzaki
}
