% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curvaturef.R
\name{get_curvaturef}
\alias{get_curvaturef}
\title{Compute the curvature between two specific points of the profile.}
\usage{
get_curvaturef(slop1, slop2)
}
\arguments{
\item{slop1}{a numeric scalar. The slope value at the first point.}

\item{slop2}{a numeric scalar. The slope value at the second point.}
}
\value{
a numeric scalar. The value of curvature between two specific points.
}
\description{
Compute the curvature between two specific points of the profile.
}
\examples{
get_curvaturef(slop1 = -1, slop2 = 1)
}
\author{
Danai Kafetzaki
}
