% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_GAM.R
\name{get_data_GAM}
\alias{get_data_GAM}
\title{Creates the data needed for computing the GAM representation}
\usage{
get_data_GAM(df_img_check, h_sc)
}
\arguments{
\item{df_img_check}{Image object, as a result of EBImage formal class Image.}

\item{h_sc}{a numeric scalar. The height for re-scaling the profile (in pixels). This should result in a global height for all profiles in the dataset.}
}
\value{
A list with the following items:
\itemize{
\item df_img_i_check - a numeric 2D matrix. The data values for the rescaled profile, ranging from zero to one.
}
}
\description{
Creates the data needed for computing the GAM representation
}
\examples{
load(file = "~/myRpacks/morphotype/data/df_img_check.RData")
get_data_GAM(df_img_check = df_img_check, h_sc = 300)
}
\author{
Danai Kafetzaki
}
