% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eccentricity.R
\name{get_eccentricity}
\alias{get_eccentricity}
\title{Compute the eccentricity of the profile segment}
\usage{
get_eccentricity(
  profile.wall,
  outer.line,
  inner.line,
  m9,
  m28,
  filename,
  writeImgA,
  writeImgwd
)
}
\arguments{
\item{profile.wall}{a numeric 2D vector. The profile binary matrix.}

\item{outer.line}{a numeric 1D vector. The outer profile line.}

\item{inner.line}{a numeric 1D vector. The inner profile line.}

\item{m9}{a numeric scalar. The height of the profile.}

\item{m28}{a numeric scalar. The height of the profile segment.}

\item{filename}{a character vector. Name of the image file including extension. The supported extensions are jpeg, png and tiff.}

\item{writeImgA}{a logical value. Whether to write the image of the profile segment. ---- MOVE TO OTHER FUNCTION ----}

\item{writeImgwd}{a character vector. Directory in which the image file is written. ---- MOVE TO OTHER FUNCTION ----}
}
\value{
A list with the following items:
\itemize{
\item eccentricity - a numeric scalar. The major axis value of the profile segment.
\item mass.centre.x - a numeric scalar. The x position of the profile segment center of mass.
\item mass.centre.y - a numeric scalar. The y position of the profile segment center of mass.
\item majoraxis - a numeric scalar. The major axis value of the profile segment.
\item df_img_rim - a numeric 2D vector. The binary matrix of the profile segment.
}
}
\description{
Compute the eccentricity of the profile segment
}
\examples{
get_eccentricity(profile.wall = m19$profile.wall, outer.line = m19$outer.line,
                 inner.line = m19$inner.line, m9 = m9, m28 = 67,
                 filename = filename, writeImgA = TRUE, writeImgwd = '~/')
}
\author{
Danai Kafetzaki
}
