% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elongation.R
\name{get_elongation}
\alias{get_elongation}
\title{Compute the eccentricity of the profile segment}
\usage{
get_elongation(m28, m47)
}
\arguments{
\item{m28}{a numeric scalar. The height of the profile segment.}

\item{m47}{a list. The result of the get_average_wall.}
}
\value{
A list with the following items:
\itemize{
\item elongation.min - a numeric scalar. The elongation using as basis the minimum wall thickness.
\item elongation.max - a numeric scalar. The elongation using as basis the maximum wall thickness.
\item elongation.avg - a numeric scalar. The elongation using as basis the average wall thickness.
}
}
\description{
Compute the eccentricity of the profile segment
}
\examples{
get_elongation(m28 = 67, m47 = m47)
}
\author{
Danai Kafetzaki
}
