% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent1.R
\name{get_extent1}
\alias{get_extent1}
\title{Compute the extent of the profile segment}
\usage{
get_extent1(profile.wall, m9, m28)
}
\arguments{
\item{profile.wall}{a numeric 2D vector. The profile binary matrix.}

\item{m9}{a numeric scalar. The height of the profile.}

\item{m28}{a numeric scalar. The height of the profile segment.}
}
\value{
a numeric scalar. The extent of the profile segment.
}
\description{
Compute the extent of the profile segment
}
\examples{
get_extent1(profile.wall = m19$profile.wall, m9 = m9, m28 = 67)
}
\author{
Danai Kafetzaki
}
