% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exterior_lines.R
\name{get_exterior_lines}
\alias{get_exterior_lines}
\title{Compute the y positions of each potential groove line}
\usage{
get_exterior_lines(image_profile, margin, sherd_rimDiam)
}
\arguments{
\item{image_profile}{a numeric 2D matrix. The binary matrix with the drawing cropped at its bounding box, as a result of get_vessel.bbox.}

\item{margin}{a numeric scalar. The margin as the number of pixels on which the potential grooves will be checked.}

\item{sherd_rimDiam}{a numeric scalar. The rim diameter of the profile in pixels.}
}
\value{
A list with the following item:
\itemize{
\item list_potGroov_ext - a numeric 1D vector. The y position of each potential groove line.
}
}
\description{
Compute the y positions of each potential groove line
}
\examples{
get_exterior_lines(image_profile = m7$image_profile, margin = 10, sherd_rimDiam = m13$sherd_rimDiam)
}
\author{
Danai Kafetzaki
}
