% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flat_rim.R
\name{get_flat_rim}
\alias{get_flat_rim}
\title{Compute how flat the top of the profile is}
\usage{
get_flat_rim(img, ym, ty, xrm)
}
\arguments{
\item{img}{a numeric 2D matrix. The binary matrix of the input image as a result of get_input.}

\item{ym}{a numeric vector. The index for each point on the y axis containing data.}

\item{ty}{a numeric scalar. The y position of the top line as a result of get_topline.}

\item{xrm}{a numeric scalar. The x position of the highest rim point as a result of get_topline.}
}
\value{
a numeric scalar. The width in pixels for the flat top of the rim.
}
\description{
Compute how flat the top of the profile is
}
\examples{
get_flat_rim(img = image_denoised, ym = m2$Y_min_i, ty = m4$topline_Yi, xrm = m4$X_Rim_min_updated)
}
\author{
Danai Kafetzaki
}
