% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_height.R
\name{get_height}
\alias{get_height}
\title{Compute the height of the profile}
\usage{
get_height(Y_rim_pos, Y_bot_pos)
}
\arguments{
\item{Y_rim_pos}{a named list of 1. The y position for the highest profile point.}

\item{Y_bot_pos}{a named list of 1. The y position for the lowest profile point.}
}
\value{
a numeric scalar. The height of the profile in pixels.
}
\description{
Compute the height of the profile
}
\examples{
get_height(Y_rim_pos = m7$bbox_TLBR[1], Y_bot_pos = m7$bbox_TLBR[3])
}
\author{
Danai Kafetzaki
}
