% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HW.R
\name{get_HW}
\alias{get_HW}
\title{Compute the ratio between height and rim diameter}
\usage{
get_HW(sherd_height, sherd_rimDiam)
}
\arguments{
\item{sherd_height}{a numeric scalar. The height of the profile.}

\item{sherd_rimDiam}{a numeric scalar. The rim diameter of the profile.}
}
\value{
a numeric scalar. The ratio of the height over the rim diameter.
}
\description{
Compute the ratio between height and rim diameter
}
\examples{
get_HW(sherd_height = m9, sherd_rimDiam = m13$sherd_rimDiam)
}
\author{
Danai Kafetzaki
}
