% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inclination_sin.R
\name{get_inclination_sin}
\alias{get_inclination_sin}
\title{Compute the inclination of the profile}
\usage{
get_inclination_sin(xx, m5)
}
\arguments{
\item{xx}{a dataframe. Melted profile matrix as a result of get_transforms.}

\item{m5}{a numeric scalar. The flatness of the rim in pixels as a result of get_flat.rim.}
}
\value{
A list with the following items:
\itemize{
\item sherd_incl_distance - a numeric scalar. Actual length in pixels of the conceivable hypotenuse for the triangle measuring sherd_incl_sin.
\item sherd_incl_sin - a numeric scalar. Sine value for the inclination measured at the profile top most inner point and the bottom most outer point.
}
}
\description{
Compute the inclination of the profile
}
\examples{
get_inclination_sin(xx = m10, m5 = m5)
}
\author{
Danai Kafetzaki
}
