% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input.R
\name{get_input}
\alias{get_input}
\title{Read input image}
\usage{
get_input(filename, trim, thr, wd)
}
\arguments{
\item{filename}{a character vector. Name of the image file including extension. The supported extensions are jpeg, png and tiff. The image should be provided in gray scale.}

\item{trim}{a numeric scalar. The pixels the image should be trimmed. This parameter should be larger than the image border. If there is no border trim can be zero.}

\item{thr}{a numeric scalar. The value of the threshold in (0,1]. A value of one means no threshold is applied. A value of zero means that all values are filtered out. A value of 0.9 is recommended.}

\item{wd}{a character vector. Directory in which the image file is read from.}
}
\value{
A list with the following items:
\itemize{
\item matrix - The 2D image matrix.
\item alternative - Whether the image should be checked.
}
}
\description{
Input a technical drawing as image file and transform it to a numerical matrix with black and white values.
}
\details{
these are details ---- CHANGE and INCLUDE in each function that needs it ----
}
\examples{
get_input(filename='SADR010324.jpg',trim=10,thr=0.9,wd='~/myRpacks/morphotype/inst/extdata')
}
\author{
Danai Kafetzaki
}
