% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_diameter.R
\name{get_max_diameter}
\alias{get_max_diameter}
\title{Compute the maximum diameter of the profile and its location}
\usage{
get_max_diameter(image_profile, xx)
}
\arguments{
\item{image_profile}{a numeric 2D matrix. The binary matrix with the drawing cropped at its bounding box, as a result of get_vessel.bbox.}

\item{xx}{a dataframe. Melted profile matrix as a result of get_transforms.}
}
\value{
A list with the following items:
\itemize{
\item max.diameter - a numeric scalar. The maximum diameter of the profile in pixels.
\item max.diameter.loc - a numeric scalar. The location of the maximum diameter from the top of the profile in pixels.
\item YposXmin - a numeric 1D vector. The y position in the image_profile of the maximum diameter on the left of the profile.
\item YposXmax - a numeric 1D vector. The y position in the image_profile of the maximum diameter on the right of the profile.
}
}
\description{
Compute the maximum diameter of the profile and its location
}
\examples{
get_max_diameter(image_profile = m7$image_profile, xx = m10)
}
\author{
Danai Kafetzaki
}
