% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MDH.R
\name{get_MDH}
\alias{get_MDH}
\title{Compute the ratio between height and maximum diameter}
\usage{
get_MDH(sherd_maxDiam_loc, sherd_height)
}
\arguments{
\item{sherd_maxDiam_loc}{a numeric scalar. The maximum diameter of the profile.}

\item{sherd_height}{a numeric scalar. The height of the profile.}
}
\value{
a numeric scalar. The ratio of the height over the maximum diameter.
}
\description{
Compute the ratio between height and maximum diameter
}
\examples{
get_MDH(sherd_maxDiam_loc = m11$max.diameter.loc, sherd_height = m9)
}
\author{
Danai Kafetzaki
}
