% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection_x.R
\name{get_projection_x}
\alias{get_projection_x}
\title{One dimensional projection of the profile on the x-axis}
\usage{
get_projection_x(image_denoised)
}
\arguments{
\item{image_denoised}{a numeric 2D matrix. The binary matrix with the drawing cropped at its bounding box, as a result of get_vessel.bbox.}
}
\value{
A list with the following items:
\itemize{
\item xx_range - a numeric 1D vector. Frequency of data per x axis index.
\item X_min - a numeric scalar. The first x axis index which contains data.
\item X_max - a numeric scalar. The last x axis index which contains data.
}
}
\description{
One dimensional projection of the profile on the x-axis
}
\examples{
get_projection_x(image_denoised = image_denoised)
}
\author{
Danai Kafetzaki
}
