% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection_y.R
\name{get_projection_y}
\alias{get_projection_y}
\title{One dimensional projection of the profile on the y-axis}
\usage{
get_projection_y(image_denoised)
}
\arguments{
\item{image_denoised}{a numeric 2D matrix. The binary matrix with the drawing cropped at its bounding box, as a result of get_vessel.bbox.}
}
\value{
A list with the following items:
\itemize{
\item yy_range - a numeric 1D vector. Frequency of data per y axis index.
\item Y_min_i - a numeric vector. The index for each point on the y axis containing data.
\item X_Rim_min - a numeric scalar.The x axis index of the top outer profile point.
}
}
\description{
One dimensional projection of the profile on the y-axis
}
\examples{
get_projection_y(image_denoised = image_denoised)
}
\author{
Danai Kafetzaki
}
