% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radius_ratio.R
\name{get_radius.ratio}
\alias{get_radius.ratio}
\title{Compute the radius ratio for a profile segment}
\usage{
get_radius.ratio(m47)
}
\arguments{
\item{m47}{a list with the following items:
\itemize{
\item all_WT - a numeric 1D vector including the wall thickness in pixels across the height of the profile wall.
\item mean.WT - the average of all_WT.
\item median.WT - the median of all_WT.
\item sd.WT - the standard deviation of all_WT.
\item min.WT - the minimum of all_WT.
\item max.WT - the maximum of all_WT.
}}
}
\value{
a numeric scalar. The radius ratio of the profile segment.
}
\description{
Compute the radius ratio for a profile segment
}
\examples{
get_radius.ratio(m47 = m47)
}
\author{
Danai Kafetzaki
}
