% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Rim_average_wall.R
\name{get_Rim_average_wall}
\alias{get_Rim_average_wall}
\title{Compute the wall thickness across the rim profile height}
\usage{
get_Rim_average_wall(outL, innL, m28)
}
\arguments{
\item{outL}{a numeric 1D vector. The outer part of the profile outline.}

\item{innL}{a numeric 1D vector. The inner part of the profile outline.}

\item{m28}{a numeric scalar. The height of the rim profile.}
}
\value{
A list with the following items:
\itemize{
\item all_WT - a numeric 1D vector including the wall thickness in pixels across the height of the profile wall.
\item mean.WT - the average of all_WT.
\item median.WT - the median of all_WT.
\item sd.WT - the standard deviation of all_WT.
\item min.WT - the minimum of all_WT.
\item max.WT - the maximum of all_WT.
}
}
\description{
Compute the wall thickness across the rim profile height
}
\examples{
get_Rim_average_wall(outL = m19$outer.line, innL = m19$inner.line, m28 = 67)
}
\author{
Danai Kafetzaki
}
