% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rim_block.R
\name{get_rim_block}
\alias{get_rim_block}
\title{Compute the percentage the rim profile resembles a block}
\usage{
get_rim_block(rim.height, sherd.height, outL, all_WT)
}
\arguments{
\item{rim.height}{a numeric scalar. The height of the rim profile.}

\item{sherd.height}{a numeric scalar. The height of the profile.}

\item{outL}{a numeric 1D vector. The outer profile line.}

\item{all_WT}{a numeric 1D vector. The wall thickness in pixels across the height of the profile wall.}
}
\value{
A list with the following item:
\itemize{
\item block.perC - a numeric scalar. The ratio of resemblance to a block.
}
}
\description{
Compute the percentage the rim profile resembles a block
}
\examples{
get_rim_block(rim.height=67, sherd.height=m9, outL=m19$outer.line, all_WT=m21$all_WT)
}
\author{
Danai Kafetzaki
}
