% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rim_manual_ggplot.R
\name{get_rim_manual_ggplot}
\alias{get_rim_manual_ggplot}
\title{Plot rim height, groove and ridge lines on profile}
\usage{
get_rim_manual_ggplot(filename, m20, m19, m28, m9)
}
\arguments{
\item{filename}{a character vector. Name of the image file.}

\item{m20}{output of get_exterior_lines.}

\item{m19}{output of get_wall_lines.}

\item{m28}{a numeric scalar. The point of rim end manual from the highest point of the profile, in pixels.}

\item{m9}{output of get_height.}
}
\value{
ggplot of the profile and annotations for the rim height, groove and ridge lines
}
\description{
Plot rim height, groove and ridge lines on profile
}
\examples{
get_rim_manual_ggplot(filename = 'SADR010324.jpg', m20 = m20, m19 = m19, m28 = 67, m9 = m9)
}
\author{
Danai Kafetzaki
}
