% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rim_measures.R
\name{get_rim_measures}
\alias{get_rim_measures}
\title{Compute rim measures of protuberance and difference length}
\usage{
get_rim_measures(m9, m28, outL, innL, all_WT)
}
\arguments{
\item{m9}{a numeric scalar. The height of the profile.}

\item{m28}{a numeric scalar. The height of the rim profile.}

\item{outL}{a numeric 1D vector. The outer profile line.}

\item{innL}{a numeric 1D vector. The inner profile line.}

\item{all_WT}{a numeric 1D vector. The wall thickness in pixels across the height of the profile wall.}
}
\value{
A list with the following items:
\itemize{
\item rim.height - a numeric scalar. The height of the rim profile.
\item rim.out.prot.length - a numeric scalar. The outer protuberance length of the rim profile.
\item rim.inn.prot.length - a numeric scalar. The inner protuberance length of the rim profile.
\item rim.out.diff.length - a numeric scalar. The outer difference length of the rim profile.
\item rim.inn.diff.length - a numeric scalar. The inner difference length of the rim profile.
\item rim.out.max.diff.length - a numeric scalar. The maximum outer difference length of the rim profile.
\item rim.inn.max.diff.length - a numeric scalar. The maximum inner difference length of the rim profile.
\item rim.out.prot.length.2 - a numeric scalar. The alternative outer protuberance length of the rim profile.
\item rim.inn.prot.length.2 - a numeric scalar. The alternative inner protuberance length of the rim profile.
}
}
\description{
Compute rim measures of protuberance and difference length
}
\examples{
get_rim_measures(m9=m9, m28=67, outL=m19$outer.line, innL=m19$inner.line, all_WT=m21$all_WT)
}
\author{
Danai Kafetzaki
}
