% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Rim_norm_wall.R
\name{get_Rim_norm_wall}
\alias{get_Rim_norm_wall}
\title{Compute the normalized wall thickness across the profile height and relevant statistics}
\usage{
get_Rim_norm_wall(df_img_rim)
}
\arguments{
\item{df_img_rim}{a numeric 2D vector. The binary matrix of the profile segment.}
}
\value{
A list with the following items:
\itemize{
\item WT.vec_norm - a numeric 1D vector. The wall thickness in pixels across the height of the normalized rim profile.
\item WT.mean_norm - a numeric scalar. The average of WT.vec_norm.
\item WT.med_norm - a numeric scalar. The median of WT.vec_norm.
\item WT.var_norm - a numeric scalar. The variance of WT.vec_norm.
\item WT.skew_norm - a numeric scalar. The skewness of WT.vec_norm.
\item WT.kurt_norm - a numeric scalar. The kurtosis of WT.vec_norm.
\item WT.BotMed_norm - a numeric scalar. The bottom width of the normalized rim profile.
}
}
\description{
Compute the normalized wall thickness across the profile height and relevant statistics
}
\examples{
get_Rim_norm_wall(df_img_rim = m52$df_img_rim)
}
\author{
Danai Kafetzaki
}
