% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rim_rounded.R
\name{get_rim_rounded}
\alias{get_rim_rounded}
\title{Compute rim measures of roundness}
\usage{
get_rim_rounded(profile.wall, sherd.height, rim.height)
}
\arguments{
\item{profile.wall}{a numeric 2D vector. The profile binary matrix.}

\item{sherd.height}{a numeric scalar. The height of the profile.}

\item{rim.height}{a numeric scalar. The height of the rim profile.}
}
\value{
A list with the following items:
\itemize{
\item FormFactor - a numeric scalar. The form factor of the profile segment.
\item AspectRatio - a numeric scalar. The aspect ratio of the profile segment.
\item Roundness - a numeric scalar. The roundness of the profile segment.
}
}
\description{
Compute rim measures of roundness
}
\examples{
get_rim_rounded(profile.wall = m19$profile.wall, sherd.height = m9, rim.height = 67)
}
\author{
Danai Kafetzaki
}
