% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rim_thickness_HC.R
\name{get_rim_thickness_HC}
\alias{get_rim_thickness_HC}
\title{Compute the horizontal cut of rim profile}
\usage{
get_rim_thickness_HC(outL, m28, innL)
}
\arguments{
\item{outL}{a numeric 1D vector. The outer profile line.}

\item{m28}{a numeric scalar. The height of the rim profile.}

\item{innL}{a numeric 1D vector. The inner profile line.}
}
\value{
A list with the following items:
\itemize{
\item Rim_HorizCut.Out_WT - a numeric scalar. The horizontal cut length at the outer of the rim profile.
\item Rim_HorizCut.Out_Width - a numeric scalar. The width of Rim_HorizCut.Out_WT.
\item Rim_HorizCut.Out_Cont - a character vector. Checks whether there are discontinuities in Rim_HorizCut.Out_WT.
\item Rim_HorizCut.Inn_WT - a numeric scalar. The horizontal cut length at the inner of the rim profile.
\item Rim_HorizCut.Inn_Width - a numeric scalar. The width of Rim_HorizCut.Inn_WT.
\item Rim_HorizCut.Inn_Cont - a character vector. Checks whether there are discontinuities in Rim_HorizCut.Inn_WT.
}
}
\description{
Compute the horizontal cut of rim profile
}
\examples{
get_rim_thickness_HC(outL = m19$outer.line, m28 = 67, innL = m19$inner.line)
}
\author{
Danai Kafetzaki
}
