% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rim_trapezoid.R
\name{get_rim_trapezoid}
\alias{get_rim_trapezoid}
\title{Compute the percentage the rim profile resembles a trapezoid}
\usage{
get_rim_trapezoid(
  rim.height,
  sherd.height,
  rim.flattness,
  rim.bottom.width,
  all_WT
)
}
\arguments{
\item{rim.height}{a numeric scalar. The height of the rim profile.}

\item{sherd.height}{a numeric scalar. The height of the profile.}

\item{rim.flattness}{a numeric scalar. The rim flatness in pixels as results from get_circumference.}

\item{rim.bottom.width}{a numeric scalar. The rim bottom width in pixels as results from get_circumference.}

\item{all_WT}{a numeric 1D vector. The wall thickness in pixels across the height of the profile wall.}
}
\value{
A list with the following item:
\itemize{
\item trapez.perC - a numeric scalar. The ratio of resemblance to a trapezoid.
}
}
\description{
Compute the percentage the rim profile resembles a trapezoid
}
\examples{
get_rim_trapezoid(rim.height = 67, sherd.height = m9, rim.flattness = m5,
                  rim.bottom.width = m38$bottom_width, all_WT = m21$all_WT)
}
\author{
Danai Kafetzaki
}
