% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scale.R
\name{get_scale}
\alias{get_scale}
\title{Compute the pixels included in the scale}
\usage{
get_scale(image_denoised, Y_min_i)
}
\arguments{
\item{image_denoised}{a numeric 2D matrix. The binary image matrix as a result of get_input.}

\item{Y_min_i}{a numeric vector. The index for each point on the y axis containing data as a result of get_projection_y.}
}
\value{
A list with the following items:
\itemize{
\item scale - a numeric scalar. The number of pixels included in the scale.
\item scale_alt - a numeric scalar. The alternative number of pixels included in the scale.
\item scaleCheck - a numeric scalar. Variance of scale.
\item Scale_Bot - a numeric scalar. The y index the scale starts.
\item scale_End - a numeric scalar. The y index the scale stops.
}
}
\description{
Compute the pixels included in the scale
}
\examples{
get_scale(image_denoised = image_denoised, Y_min_i = m2$Y_min_i)
}
\author{
Danai Kafetzaki
}
