% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scaled_rim_outlines.R
\name{get_scaled_rim_outlines}
\alias{get_scaled_rim_outlines}
\title{Scale the profile and retrieves the outline}
\usage{
get_scaled_rim_outlines(profile.wall, inner.line, m9, m28, hsc)
}
\arguments{
\item{profile.wall}{a numeric 2D vector. The profile binary matrix.}

\item{inner.line}{a numeric 1D vector. The inner profile line.}

\item{m9}{a numeric scalar. The height of the profile.}

\item{m28}{a numeric scalar. The height of the rim profile.}

\item{hsc}{a numeric scalar. The new profile height in pixels.}
}
\value{
A list with the following items:
\itemize{
\item df.img.rim.scaled - a numeric 2D vector. The scaled profile binary matrix.
\item outer.rim.line.scaled - a numeric 1D vector. The outer line of the scaled profile.
\item inner.rim.line.scaled - a numeric 1D vector. The inner line of the scaled profile.
}
}
\description{
Scale the profile and retrieves the outline
}
\examples{
get_scaled_rim_outlines(profile.wall = m19$profile.wall,
                        inner.line = m19$inner.line,
                        m9 = m9, m28 = 67, hsc = 90)
}
\author{
Danai Kafetzaki
}
