% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoothed_lines.R
\name{get_smoothed_lines}
\alias{get_smoothed_lines}
\title{Compute the smooth outlines of the profile}
\usage{
get_smoothed_lines(data_profile)
}
\arguments{
\item{data_profile}{a numeric 2D vector. The profile binary matrix.}
}
\value{
A list with the following items:
\itemize{
\item data_circumference - a numeric 2D vector. The circumference of the profile.
\item data_circumference_eroded - a numeric 2D vector. The eroded circumference of the profile.
\item outer.line.mod - a numeric 2D vector. The smoothed outer profile line.
\item outer.line.mod.cor - a numeric 2D vector. The corrected smoothed outer profile line.
\item bottom.line.mod - a numeric 2D vector. The smoothed bottom outer profile line.
\item bottom.line.mod.cor - a numeric 2D vector. The corrected smoothed bottom outer profile line.
\item fit_splines - an object of class "smooth.spline". The spline model for outer.line.mod.cor fitted with cross-validated knots.
\item fit_splines_max - an object of class "smooth.spline". The spline model for outer.line.mod.cor fitted with maximum knots.
\item fit_splines_bot - an object of class "smooth.spline". The spline model for bottom.line.mod.cor fitted with cross-validated knots.
\item fit_splines_max_bot - an object of class "smooth.spline". The spline model for bottom.line.mod.cor fitted with maximum knots.
\item df_bot - a melted dataframe. The melted data_circumference_eroded.
\item fit_splines.inn - an object of class "smooth.spline". The spline model for df_mod.inn_cor fitted with cross-validated knots.
\item fit_splines.inn_max - an object of class "smooth.spline". The spline model for df_mod.inn_cor fitted with maximum knots.
\item fit_splines.inn_bot - an object of class "smooth.spline". The spline model for df_mod.inn_bot_cor fitted with cross-validated knots.
\item fit_splines.inn_max_bot - an object of class "smooth.spline". The spline model for df_mod.inn_bot_cor fitted with maximum knots.
\item df_mod.inn_cor - a numeric 2D vector. The corrected smoothed inner profile line.
\item df_mod.inn_bot_cor - a numeric 2D vector. The corrected smoothed bottom inner profile line.
}
}
\description{
Compute the smooth outlines of the profile
}
\examples{
get_smoothed_lines(data_profile = m19$profile.wall)
}
\author{
Danai Kafetzaki
}
