% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tip_thickness.R
\name{get_tip_thickness}
\alias{get_tip_thickness}
\title{Compute the thickness at the rim tip}
\usage{
get_tip_thickness(image_denoised, X_min)
}
\arguments{
\item{image_denoised}{a numeric 2D vector. The profile binary matrix.}

\item{X_min}{a numeric scalar. The first x axis index which contains data.}
}
\value{
a numeric scalar. The width in pixels for the flat top of the rim.
}
\description{
Compute the thickness at the rim tip
}
\examples{
get_tip_thickness(image_denoised = image_denoised, X_min = m1$X_min)
}
\author{
Danai Kafetzaki
}
