% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topline.R
\name{get_topline}
\alias{get_topline}
\title{Compute position of the profile rim line}
\usage{
get_topline(image_denoised, Y_min_i, margin)
}
\arguments{
\item{image_denoised}{Output of get_input.}

\item{Y_min_i}{Output of get_projection_y.}

\item{margin}{Numeric value. The range in pixels the topline is checked.}
}
\value{
A list with the following items:
\itemize{
\item topline_Yi - a numeric scalar. The y axis distance in pixels from the first data point to the rim line.
\item X_Rim_min_updated - a numeric scalar. The updated y axis index for the rim line.
}
}
\description{
Compute position of the profile rim line
}
\examples{
get_topline(image_denoised = image_denoised, Y_min_i = m2$Y_min_i, margin = 12)
}
\author{
Danai Kafetzaki
}
