% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vec_df.R
\name{get_vec_df}
\alias{get_vec_df}
\title{Return a list of dataframes with the profile planar curves}
\usage{
get_vec_df(m19, m32o, m32i, m33, m34o, m34i, m5, m9, mY, fl)
}
\arguments{
\item{m19}{a list. Output of get_wall_lines.}

\item{m32o}{a list. Output of get_curvature for the outer rim profile line.}

\item{m32i}{a list. Output of get_curvature for the inner rim profile line.}

\item{m33}{a list. Output of get_smoothed.lines for the profile binary matrix.}

\item{m34o}{a list. Output of get_curvature for the outer rim profile splines model.}

\item{m34i}{a list. Output of get_curvature for the inner rim profile splines model.}

\item{m5}{a numeric scalar. The width in pixels for the flat top of the rim.}

\item{m9}{a numeric scalar. The height of the profile.}

\item{mY}{a numeric scalar. The height of the profile minus the height of the rim profile.}

\item{fl}{a character vector. Name of the image file including extension. The supported extensions are jpeg, png and tiff.}
}
\value{
A list with the following dataframes:
\itemize{
\item perip.original - The original periphery data.
\item perip.smoothed - The smoothed periphery data.
\item slop.original - The original slope data.
\item slop.smoothed - The smoothed slope data.
\item curvf.original - The original curvature data.
\item curvf.smoothed - The smoothed curvature data.
}
}
\description{
Return a list of dataframes with the profile planar curves
}
\examples{
 get_vec_df(m19 = m19, m32o = m32o, m32i = m32i, m33 = m33,
            m34o = m34o, m34i = m34i, m5 = m5, m9 = m9,
            mY = mlypos, fl = filename)
}
\author{
Danai Kafetzaki
}
