% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessel_bbox.R
\name{get_vessel_bbox}
\alias{get_vessel_bbox}
\title{Compute the bounding box of the profile}
\usage{
get_vessel_bbox(img, ym, yt, yb, m1)
}
\arguments{
\item{img}{a numeric 2D matrix. The binary matrix of the input image as a result of get_input.}

\item{ym}{a numeric vector. The index for each point on the y axis containing data.}

\item{yt}{a numeric scalar. The y axis distance in pixels from the first data point to the rim line.}

\item{yb}{a numeric scalar. The y coordinate of the most bottom point of the profile.}

\item{m1}{a list of 3. Output of get_projection_x.}
}
\value{
A list with the follwoing items:
\itemize{
\item image_profile - a numeric 2D matrix. The binary matrix with the drawing cropped at its bounding box, as a result of get_vessel.bbox.
\item bbox_TLBR - a list of four numeric scalars. The bounding box of the profile drawing.
}
}
\description{
Compute the bounding box of the profile
}
\examples{
get_vessel_bbox(img=image_denoised, ym=m2$Y_min_i, yt=m4$topline_Yi, yb=m6, m1=m1)
}
\author{
Danai Kafetzaki
}
