% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wall_lines.R
\name{get_wall_lines}
\alias{get_wall_lines}
\title{Return the main wall lines of the profile}
\usage{
get_wall_lines(img, m, mm, rd, m9, filenaam)
}
\arguments{
\item{img}{a numeric 2D matrix. The binary matrix with the drawing cropped at its bounding box, as a result of get_vessel.bbox.}

\item{m}{a numeric scalar. The number of pixels checked for correction in the calculation.}

\item{mm}{a numeric scalar. The amount of pixels allowed to be missing in the potential groove line such that grooves that are not drawn to reach correctly the middle line are included.}

\item{rd}{a numeric scalar. The rim diameter.}

\item{m9}{a numeric scalar. The height of the profile in pixels.}

\item{filenaam}{a character vector. Name of the image file including extension. The supported extensions are jpeg, png and tiff.}
}
\value{
A list with the following items:
\itemize{
\item profile.wall - a numeric 2D matrix. The binary matrix of the profile wall.
\item outer.line - a numeric 1D vector. The outer profile line.
\item inner.line - a numeric 1D vector. The inner profile line.
\item list_potGroov_in - a character vector. The y index of the potential exterior grooves separated by comma.
\item listlines_in - a character vector. The y index of the potential interior grooves separated by comma.
\item list_corr - a character vector. The y index of all lines in the profile reaching the midline separated by comma.
\item has.fullbase - a Boolean. Whether the profile has complete base.
\item fullbase - a character vector. The profiles y index of the base beginning and end separated by a colon.
\item base.height - a numeric scalar. The height of the base in pixels.
\item is.full.profile - a Boolean. Whether the profile is complete.
}
}
\description{
Return the main wall lines of the profile
}
\examples{
get_wall_lines(img=m7$image_profile,m=10,mm=3,rd=m13$sherd_rimDiam,m9=m9,filenaam='SADR010324.jpg')
}
\author{
Danai Kafetzaki
}
