% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wall_thickness.R
\name{get_wall_thickness}
\alias{get_wall_thickness}
\title{Compute the wall thickness at a specific location}
\usage{
get_wall_thickness(xx, loc)
}
\arguments{
\item{xx}{a dataframe. Melted profile matrix as a result of get_transforms.}

\item{loc}{a numeric scalar. The location in pixels where the wall thickness should be computed. It takes minimum value 1 and maximum equal to the profile height.}
}
\value{
a numeric scalar. The wall thickness in pixels at a specific location.
}
\description{
Compute the wall thickness at a specific location
}
\examples{
get_wall_thickness(xx = m10, loc = round(2*m9/3))
}
\author{
Danai Kafetzaki
}
