% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_denoised.R
\docType{data}
\name{image_denoised}
\alias{image_denoised}
\title{Result of get_input.R}
\format{
Large numeric matrix equal to the size of the original technical drawing.
\describe{
Rows correspond to the width of the technical drawing.
Columns correspond to the height of the technical darwing.
}
}
\source{
\url{https://www.dataURL.com}
}
\usage{
image_denoised
}
\description{
The image transformed in a matrix containing zero and one values.
}
\keyword{datasets}
