% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1.R
\docType{data}
\name{m1}
\alias{m1}
\title{Output of get_projection_x}
\format{
A list with 3 items:
\describe{
\item{xx_range}{description}
\item{X_min}{description}
\item{X_max}{description}
}
}
\source{
\url{https://www.dataURL.com}
}
\usage{
m1
}
\description{
More about Output of get_projection_x
}
\keyword{datasets}
