% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m10.R
\docType{data}
\name{m10}
\alias{m10}
\title{Output of get_transforms}
\format{
Melted data of the profile.
\describe{
\item{Var1}{description}
\item{Var2}{description}
\item{value}{description}
}
}
\source{
\url{https://www.dataURL.com}
}
\usage{
m10
}
\description{
More about Output of get_transforms / Melted data of the profile.
}
\keyword{datasets}
