% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_data}
\alias{map_data}
\title{Estimate graph from distance matrix}
\usage{
map_data(distM, filter_values = NULL, num_intervals = NULL)
}
\arguments{
\item{distM}{a distance matrix. Computed using compute_dist_0, compute_dist_1 or compute_dist_2.}

\item{filter_values}{stad parameter. Default is NULL.}

\item{num_intervals}{stad parameter. Default is NULL.}
}
\value{
A list with the following items:
\itemize{
\item graph_est - an igraph object. The estimated graph which is either the graps estimated based on the stad algorithm or the minimum spanning tree.
\item df_links - a dataframe. Contains the links of the graph under the columns 'Source' and 'Target' and the edge weight under name 'Value2'.
\item plot.shepard - a list of four. Shepard diagram data as provided by MASS::Shepard plus the Pearson correlation value as quality measure for the map estimation.
}
}
\description{
Estimate graph from distance matrix
}
\examples{
distM = read.csv('~/myRpacks/morphotype/inst/extdata/distM.csv', row.names = 1)
map_data(distM)
}
\author{
Danai Kafetzaki
}
