% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_circumference_scaled.R
\docType{data}
\name{mat_circumference_scaled}
\alias{mat_circumference_scaled}
\title{Data of ....}
\format{
A dataframe with the following variables:
\describe{
\item{Var1}{description}
\item{Var2}{description}
\item{Var3}{description}
}
}
\source{
\url{https://www.dataURL.com}
}
\usage{
mat_circumference_scaled
}
\description{
More about this data ...
}
\keyword{datasets}
