% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_cropped_profile_ScH.R
\docType{data}
\name{mat_cropped_profile_ScH}
\alias{mat_cropped_profile_ScH}
\title{Output of ....}
\format{
A list of 3 items:
\describe{
\item{Item1}{description}
\item{Item2}{description}
\item{Item3}{description}
}
}
\source{
\url{https://www.dataURL.com}
}
\usage{
mat_cropped_profile_ScH
}
\description{
More about Output of ...
}
\keyword{datasets}
