% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_logistic.R
\name{mf_logistic}
\alias{mf_logistic}
\title{Define high and low fuzzy sets}
\usage{
mf_logistic(indMin, indMax, seqstep, kval, x0val)
}
\arguments{
\item{indMin}{a numeric scalar. The global minimum for the universe of discourse.}

\item{indMax}{a numeric scalar. The global maximum for the universe of discourse.}

\item{seqstep}{a numeric 1D vector. Increment of the sequence in the universe of discourse.}

\item{kval}{a numeric scalar. The logistic growth rate or steepness of the curve.}

\item{x0val}{a numeric scalar. The x value of the sigmoid midpoint.}
}
\value{
A list with the following items:
\itemize{
\item xvals_mf - a numeric 1D vector. The values defining the universe of discourse.
\item yvals_mf_A - a numeric 1D vector. The values defining the fuzzy set high.
\item yvals_mf_B - a numeric 1D vector. The values defining the fuzzy set low.
}
}
\description{
Define high and low fuzzy sets
}
\examples{
mf_logistic(indMin = 0,indMax = 1, seqstep = 0.001, kval = 15, x0val = 0.5)
}
\author{
Danai Kafetzaki
}
