% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_logistic_gaussian.R
\name{mf_logistic_gaussian}
\alias{mf_logistic_gaussian}
\title{Define positive zero and negative fuzzy sets}
\usage{
mf_logistic_gaussian(indMin, indMax, seqstep, kval, x0val, mean0, sd0)
}
\arguments{
\item{indMin}{a numeric scalar. The global minimum for the universe of discourse.}

\item{indMax}{a numeric scalar. The global maximum for the universe of discourse.}

\item{seqstep}{a numeric 1D vector. Increment of the sequence in the universe of discourse.}

\item{kval}{a numeric scalar. The logistic growth rate or steepness of the curve.}

\item{x0val}{a numeric scalar. The x value of the sigmoid midpoint.}

\item{mean0}{a numeric scalar.}

\item{sd0}{a numeric scalar.}
}
\value{
A list with the following items:
\itemize{
\item xvals_mf - a numeric 1D vector. The values defining the universe of discourse.
\item yvals_mf_A - a numeric 1D vector. The values defining the fuzzy set positive.
\item yvals_mf_B - a numeric 1D vector. The values defining the fuzzy set zero.
\item yvals_mf_C - a numeric 1D vector. The values defining the fuzzy set negative.
}
}
\description{
Define positive zero and negative fuzzy sets
}
\examples{
mf_logistic_gaussian(indMin=-50,indMax=50,seqstep=0.1,kval=0.3,x0val=0,mean0=0,sd0=5)
}
\author{
Danai Kafetzaki
}
