% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_plot.R
\name{mf_plot}
\alias{mf_plot}
\title{Plot membership functions for high and low fuzzy sets}
\usage{
mf_plot(dataplot_mfs)
}
\arguments{
\item{dataplot_mfs}{a dataframe. The input data for the plot, with at least three variables: "xvals_mf", "yvals_mf" and "mf".}
}
\value{
a ggplot. The membership functions for high and low fuzzy sets.
}
\description{
Plot membership functions for high and low fuzzy sets
}
\examples{
mfl = mf_logistic(indMin = 0,indMax = 1, seqstep = 0.001, kval = 15, x0val = 0.5)
dataplot_mf_A=data.frame("xvals_mf"=mfl$xvals_mf,"yvals_mf"=mfl$yvals_mf_A,"mf"="high")
dataplot_mf_B=data.frame("xvals_mf"=mfl$xvals_mf,"yvals_mf"=mfl$yvals_mf_B,"mf"="low")
dataplot_mfs = rbind.data.frame(dataplot_mf_A, dataplot_mf_B)
mf_plot(dataplot_mfs = dataplot_mfs)

mfl=mf_logistic_gaussian(indMin=-50,indMax=50,seqstep=0.1,kval=0.3,x0val=0,mean0=0,sd0=5)
dataplot_mf_A=data.frame("xvals_mf"=mfl$xvals_mf,"yvals_mf"=mfl$yvals_mf_A,"mf"="positive")
dataplot_mf_B=data.frame("xvals_mf"=mfl$xvals_mf,"yvals_mf"=mfl$yvals_mf_B,"mf"="zero")
dataplot_mf_C=data.frame("xvals_mf"=mfl$xvals_mf,"yvals_mf"=mfl$yvals_mf_C,"mf"="negative")
dataplot_mfs=rbind.data.frame(dataplot_mf_A, dataplot_mf_B, dataplot_mf_C)
}
\author{
Danai Kafetzaki
}
