% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_GAM.R
\name{normalize_GAM}
\alias{normalize_GAM}
\title{Normalizes GAM representation}
\usage{
normalize_GAM(mat_spl_FP)
}
\arguments{
\item{mat_spl_FP}{\itemize{
\item a list with the GAM representation for each profile in the dataset.
}}
}
\value{
A list with the following items:
\itemize{
\item mat_spl_FP_sc - a list with the scaled GAM representation for each profile in the dataset. The scaling is performed such that both ends of the GAM vector are superimposed in all profiles.
}
}
\description{
Normalizes GAM representation
}
\examples{
load(file = "~/myRpacks/morphotype/data/mat_spl_FP.RData")
normalize_GAM(mat_spl_FP = mat_spl_FP)
}
\author{
Danai Kafetzaki
}
