% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GAM.R
\name{plot_GAM}
\alias{plot_GAM}
\title{Plots all GAM outlines}
\usage{
plot_GAM(mat_spl_FP, ylim_range)
}
\arguments{
\item{mat_spl_FP}{a list. The list contains the GAM representation for each profile in the dataset. Each item of the list has two variables in the following order: 'spl_x', 'spl_y', as a result of function compute_GAM().}

\item{ylim_range}{a numeric vector. The vector contains the minimum and maximum values for the y-axis.}
}
\value{
A plot of all GAM outlines, original or normalized.
}
\description{
Plots all GAM outlines
}
\examples{
load(file = "~/myRpacks/morphotype/data/mat_spl_FP.RData")
plot_GAM(mat_spl_FP = mat_spl_FP, ylim_range = c(0, 550))
}
\author{
Danai Kafetzaki
}
