% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_periphery.R
\name{scale_periphery}
\alias{scale_periphery}
\title{Scale the rim periphery and compute scaling metadata}
\usage{
scale_periphery(fl, img, m9, m19, m28, m38, pscl)
}
\arguments{
\item{fl}{a character vector. Name of the image file including extension. The supported extensions are jpeg, png and tiff.}

\item{img}{a numeric 2D matrix. The binary image matrix as a result of get_input.}

\item{m9}{a numeric scalar. The height of the profile.}

\item{m19}{a list of 10. Output of get_wall_lines.}

\item{m28}{a numeric scalar. The height of the rim profile.}

\item{m38}{a list of 3. Output of get_circumference.}

\item{pscl}{a numeric scalar. The targeted length of the periphery after scaling.}
}
\value{
A list with the following items:
\itemize{
\item df_scaled_1D.outline - a dataframe.
\item df_scaled_1D.outline.plot - a dataframe.
\item scaled_1D_fpl - a numeric scalar. The scale factor.
\item scaled_1D_height - a numeric scalar. The scaled rim height.
\item pixel_sc_dif - a numeric scalar. Difference in targeted and actual length of the scaled periphery.
\item scaled_1D_bottom_width - a numeric scalar. The width of the bottom in the scaled rim profile.
\item scaled_1D_lip_width - a numeric scalar. The width of the top in the scaled rim profile.
}
}
\description{
Scale the rim periphery and compute scaling metadata
}
\examples{
scale_periphery(fl='SADR010324.jpg',img=image_denoised,m9=m9,m19=m19,m28=67,m38=m38,pscl=217)
}
